backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
psycopg[binary]==3.1.12
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.4.0
requests==2.32.3
openai==2.18.0
httpx==0.27.2
python-docx==1.1.0
reportlab==4.4.10
gunicorn==21.2.0
pypdf==4.3.1
pdf2image==1.17.0
pdfplumber==0.11.9
google-generativeai==0.5.4

# SAT Verification API dependencies
fastapi
uvicorn
psycopg2-binary
pandas
requests
